1重构了命令处理模块 2重构了网络处理模块 两个模块彻底解耦合
