+++ RemoteCtr/RemoteCtr/Command.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "resource.h"
 #include<map>
-#include "ServerSocket.h"
+#include "Packet.h"
 #include<atlimage.h>
 #include<direct.h>
 #include "MyTool.h"
@@ -15,9 +15,23 @@ class CCommand
 public:
 	CCommand();
 	~CCommand(){}
-	int ExcuteCommand(int nCmd);
+	int ExcuteCommand(int nCmd, std::list<CPacket>& listPacket,CPacket&inPacket);
+    static void RunCommand(void* arg, int status, std::list<CPacket>&listPacket, 
+        CPacket& inPacket) {
+        CCommand* thiz = (CCommand*)arg;
+        if (status > 0) {
+            int ret = thiz->ExcuteCommand(status,listPacket,inPacket);
+            if (ret != 0) {
+                TRACE(_T("执行命令失败：%d ret=%d\r\n",status, ret));
+            }
+        }
+        else {
+            MessageBox(NULL, _T("无法正常接入用户,自动重试"),
+                _T("接入用户失败"), MB_OK | MB_ICONERROR);
+        }
+    }
 protected:
-	typedef int(CCommand::* CMDFUNC)();
+	typedef int(CCommand::* CMDFUNC)(std::list<CPacket>& listPacket, CPacket& inPacket);
 	std::map<int, CMDFUNC>m_mapFunction;
     CLockDialog dlg;
     unsigned threadid;
@@ -77,7 +91,7 @@ protected:
         ::ShowWindow(::FindWindowW(_T("Shell_TrayWnd"), NULL), SW_SHOW);
         dlg.DestroyWindow();
     }
-    int MakeDriverInfo() {
+    int MakeDriverInfo(std::list<CPacket>& listPacket, CPacket& inPacket) {
         std::string result;
         for (int i = 1; i <= 26; i++) {
             if (_chdrive(i) == 0) {
@@ -88,21 +102,18 @@ protected:
         }
         CPacket pack(1, (BYTE*)result.c_str(), result.size());//打包用的；
         CMyTool::Dump((BYTE*)pack.Data(), pack.Size());
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
-    int MakeDirectoryInfo() {
-        std::string strPath;
+    int MakeDirectoryInfo(std::list<CPacket>& listPacket, CPacket& inPacket) {
+        std::string strPath=inPacket.strData;
         //std::list<FILEINFO>listFileInfos;
-        if (CServerSocket::getInstance()->GetFilePath(strPath) == false) {
-            OutputDebugString(_T("当前命令不是获取文件列表信息，解析错误"));
-            return -1;
-        }
+        
         if (_chdir(strPath.c_str()) == -1) {
             FILEINFO finfo;
             finfo.hasNext = FALSE;
             CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-            CServerSocket::getInstance()->Send(pack);
+            listPacket.push_back(pack);
             OutputDebugString(_T("没有访问权限！！"));
             return -2;
         }
@@ -113,7 +124,7 @@ protected:
             FILEINFO finfo;
             finfo.hasNext = FALSE;
             CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-            CServerSocket::getInstance()->Send(pack);
+            listPacket.push_back(pack);
             return -3;
         }
         int count = 0;
@@ -123,7 +134,7 @@ protected:
             memcpy(finfo.szFileName, fdata.name, strlen(fdata.name));
             TRACE(("finfo.szFileName:%s\r\n"), finfo.szFileName);
             CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-            CServerSocket::getInstance()->Send(pack);
+            listPacket.push_back(pack);
             count++;
             //listFileInfos.push_back(finfo);
         } while (!_findnext(hfind, &fdata));
@@ -131,148 +142,141 @@ protected:
         FILEINFO finfo;
         finfo.hasNext = FALSE;
         CPacket pack(2, (BYTE*)&finfo, sizeof(finfo));
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
 
         return 0;
     }
 
-    int Runfile() {
-        std::string strPath;
-        CServerSocket::getInstance()->GetFilePath(strPath);
+    int Runfile(std::list<CPacket>& listPacket, CPacket& inPacket) {
+        std::string strPath = inPacket.strData;
         ShellExecuteA(NULL, NULL, strPath.c_str(), NULL, NULL, SW_SHOWNORMAL);
         CPacket pack(3, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
-    int DownloadFile() {
-        std::string strPath;
-        CServerSocket::getInstance()->GetFilePath(strPath);
+    int DownloadFile(std::list<CPacket>& listPacket, CPacket& inPacket) {
+        std::string strPath = inPacket.strData;
         FILE* file = NULL;
         errno_t err = fopen_s(&file, strPath.c_str(), "rb");
         long long data = 0;
         if (err != 0) {
             CPacket pack(4, (BYTE*)&data, 8);
-            CServerSocket::getInstance()->Send(pack);
+            listPacket.push_back(pack);
             return -1;
         }
         if (file != NULL) {
             fseek(file, 0, SEEK_END);
             data = _ftelli64(file);
             CPacket head(4, (BYTE*)&data, 8);
-            CServerSocket::getInstance()->Send(head);
+            listPacket.push_back(head);
             fseek(file, 0, SEEK_SET);
             char buff[1024] = "";
             size_t rlen = 0;
             do {
                 rlen = fread(buff, 1, 1024, file);
                 CPacket pack(4, (BYTE*)buff, rlen);
-                CServerSocket::getInstance()->Send(pack);
+                listPacket.push_back(pack);
             } while (rlen >= 1024);
             fclose(file);
         }
         CPacket pack(4, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
 
-    int MouseEvent() {
+    int MouseEvent(std::list<CPacket>& listPacket, CPacket& inPacket) {
         MOUSEEV mouse;
-        if (CServerSocket::getInstance()->GetMouseEvent(mouse)) {
-            DWORD nFlags = 0;
-            switch (mouse.nButton)
-            {
-            case 0://左键；
-                nFlags = 1;
-                break;
-            case 1://右键
-                nFlags = 2;
-                break;
-            case 2://中键；
-                nFlags = 4;
-                break;
-            case 4://没有按键，鼠标移动
-                nFlags = 8;
-                break;
-            }
-            if (nFlags != 8) {
-                SetCursorPos(mouse.ptXY.x, mouse.ptXY.y);
-            }
-            switch (mouse.nAction)
-            {
-            case 0://单击；1,2,4,8表示4个比特位；
-                nFlags |= 0x10;
-                break;
-            case 1://双击；
-                nFlags |= 0x20;
-                break;
-            case 2://按下；
-                nFlags |= 0x40;
-                break;
-            case 3://放开；
-                nFlags |= 0x80;
-                break;
-            default:
-                break;
-            }
-            TRACE("mouse event : %08X x =%d, y =%d\r\n", nFlags, mouse.ptXY.x, mouse.ptXY.y);
-            switch (nFlags)
-            {
-            case 0x21://左键双击
-                mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-            case 0x11://左键单击
-                mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x41://左键按下
-                mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x81://左键放开
-                mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-
-            case 0x22:
-                mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-            case 0x12:
-                mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x42:
-                mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x82:
-                mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-
-            case 0x24:
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-            case 0x14:
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
-                mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x44:
-                mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x84:
-                mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
-                break;
-            case 0x08:  //移动鼠标；
-                mouse_event(MOUSEEVENTF_MOVE, mouse.ptXY.x, mouse.ptXY.y, 0, GetMessageExtraInfo());
-                break;
-            }
-            CPacket pack(4, NULL, 0);
-            CServerSocket::getInstance()->Send(pack);
+        memcpy(&mouse, inPacket.strData.c_str(), inPacket.strData.size());
+        DWORD nFlags = 0;
+        switch (mouse.nButton)
+        {
+        case 0://左键；
+            nFlags = 1;
+            break;
+        case 1://右键
+            nFlags = 2;
+            break;
+        case 2://中键；
+            nFlags = 4;
+            break;
+        case 4://没有按键，鼠标移动
+            nFlags = 8;
+            break;
         }
-        else {
-            OutputDebugString(_T("获取鼠标操作参数失败！！"));
-            return -1;
+        if (nFlags != 8) {
+            SetCursorPos(mouse.ptXY.x, mouse.ptXY.y);
         }
+        switch (mouse.nAction)
+        {
+        case 0://单击；1,2,4,8表示4个比特位；
+            nFlags |= 0x10;
+            break;
+        case 1://双击；
+            nFlags |= 0x20;
+            break;
+        case 2://按下；
+            nFlags |= 0x40;
+            break;
+        case 3://放开；
+            nFlags |= 0x80;
+            break;
+        default:
+            break;
+        }
+        TRACE("mouse event : %08X x =%d, y =%d\r\n", nFlags, mouse.ptXY.x, mouse.ptXY.y);
+        switch (nFlags)
+        {
+        case 0x21://左键双击
+            mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+        case 0x11://左键单击
+            mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x41://左键按下
+            mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x81://左键放开
+            mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+
+        case 0x22:
+            mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+        case 0x12:
+            mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x42:
+            mouse_event(MOUSEEVENTF_RIGHTDOWN, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x82:
+            mouse_event(MOUSEEVENTF_RIGHTUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+
+        case 0x24:
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+        case 0x14:
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
+            mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x44:
+            mouse_event(MOUSEEVENTF_MIDDLEDOWN, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x84:
+            mouse_event(MOUSEEVENTF_MIDDLEUP, 0, 0, 0, GetMessageExtraInfo());
+            break;
+        case 0x08:  //移动鼠标；
+            mouse_event(MOUSEEVENTF_MOVE, mouse.ptXY.x, mouse.ptXY.y, 0, GetMessageExtraInfo());
+            break;
+        }
+        CPacket pack(5, NULL, 0);
+        listPacket.push_back(pack);
         return 0;
     }
 
-    int SendScreen() {
+    int SendScreen(std::list<CPacket>& listPacket, CPacket& inPacket) {
         CImage screen;//GDI
         HDC hScreen = ::GetDC(NULL);
         int nBitPerPixel = GetDeviceCaps(hScreen, BITSPIXEL);
@@ -293,7 +297,7 @@ protected:
             PBYTE pData = (PBYTE)GlobalLock(hMem);
             SIZE_T nSize = GlobalSize(hMem);
             CPacket pack(6, pData, nSize);
-            CServerSocket::getInstance()->Send(pack);
+            listPacket.push_back(pack);
             GlobalUnlock(hMem);
         }
         //screen.Save(_T("test2020.webp"), Gdiplus::ImageFormatTIFF);
@@ -302,39 +306,38 @@ protected:
         screen.ReleaseDC();
         return 0;
     }
-    int LockMachine() {
+    int LockMachine(std::list<CPacket>& listPacket, CPacket& inPacket) {
         if ((dlg.m_hWnd == NULL) || (dlg.m_hWnd == INVALID_HANDLE_VALUE)) {
             // _beginthread(threadLockDlg, 0, NULL);
             _beginthreadex(NULL, 0, threadLockDlg, this, 0, &threadid);
             TRACE("threadid=%d\r\n", threadid);
         }
         CPacket pack(7, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
 
-    int UnloakMachine() {
+    int UnloakMachine(std::list<CPacket>& listPacket, CPacket& inPacket) {
         PostThreadMessage(threadid, WM_KEYDOWN, 0x1b, 0x10001);//根据线程来的；
         CPacket pack(8, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
 
-    int TestConnect() {
+    int TestConnect(std::list<CPacket>& listPacket, CPacket& inPacket) {
         CPacket pack(1981, NULL, 0);
-        CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
 
-    int DeleteLocalFile() {
-        std::string strPath;
-        CServerSocket::getInstance()->GetFilePath(strPath);
+    int DeleteLocalFile(std::list<CPacket>& listPacket, CPacket& inPacket) {
+        std::string strPath=inPacket.strData;
         std::wstring wPath(strPath.begin(), strPath.end());
         if (DeleteFile(wPath.c_str()) == 0) {
             AfxMessageBox(_T("删除文件失败！！"));
         }
         CPacket pack(9, NULL, 0);
-        bool ret = CServerSocket::getInstance()->Send(pack);
+        listPacket.push_back(pack);
         return 0;
     }
 };
