+++ RemoteCtr/RemoteCtr/Packet.h
@@ -1,0 +1,127 @@
+#pragma once
+#include "pch.h"
+#include "framework.h"
+
+#pragma pack(push)
+#pragma pack(1)	//设置结构体、联合体和类成员的对齐方式为1字节对齐。
+
+class CPacket {
+public:
+	CPacket() :sHead(0), nLength(0), sCmd(0), sSum(0) {}
+	CPacket(WORD nCmd, const BYTE* pData, size_t nSize) {
+		sHead = 0xFEFF;
+		nLength = nSize + 4;
+		sCmd = nCmd;
+		if (nSize > 0) {
+			strData.resize(nSize);
+			memcpy((void*)strData.c_str(), pData, nSize);
+		}
+		else {
+			strData.clear();
+		}
+
+		sSum = 0;
+		for (size_t j = 0; j < strData.size(); j++) {
+			sSum += BYTE(strData[j]) & 0xFF;
+		}
+	}
+	CPacket(const CPacket& pack) {
+		sHead = pack.sHead;
+		nLength = pack.nLength;
+		sCmd = pack.sCmd;
+		strData = pack.strData;
+		sSum = pack.sSum;
+	}
+	CPacket(const BYTE* pData, size_t& nSize) {
+		size_t i = 0;
+		for (; i < nSize; i++) {
+			if (*(WORD*)(pData + i) == 0xFEFF) {
+				sHead = *(WORD*)(pData + i);
+				i += 2;
+				break;
+			}
+		}
+		if (i + 4 + 2 + 2 > nSize) {
+			nSize = 0;
+			return;
+		}
+		nLength = *(DWORD*)(pData + i); i += 4;
+		if (nLength + i > nSize) {	//包未完全接受到;
+			nSize = 0;
+			return;
+		}
+		sCmd = *(WORD*)(pData + i); i += 2;
+		if (nLength > 4) {
+			strData.resize(nLength - 4);
+			memcpy((void*)strData.c_str(), pData + i, nLength - 4);
+			i += nLength - 4;
+		}
+		sSum = *(WORD*)(pData + i); i += 2;
+		WORD sum = 0;
+		for (size_t j = 0; j < strData.size(); j++) {
+			sum += BYTE(strData[j]) & 0xFF;
+		}
+		if (sum == sSum) {
+			nSize = i;//实际使用的size；
+			return;
+		}
+		nSize = 0;
+	}
+	~CPacket() {}
+	CPacket& operator=(const CPacket& pack) {	//const
+		if (this != &pack) {
+			sHead = pack.sHead;
+			nLength = pack.nLength;
+			sCmd = pack.sCmd;
+			strData = pack.strData;
+			sSum = pack.sSum;
+		}
+		return *this;
+	}
+	int Size() {
+		return nLength + 6;
+	}
+	const char* Data() {
+		strOut.resize(nLength + 6);
+		BYTE* pData = (BYTE*)strOut.c_str();
+		*(WORD*)pData = sHead; pData += 2;
+		*(DWORD*)pData = nLength; pData += 4;
+		*(WORD*)pData = sCmd; pData += 2;
+		memcpy(pData, strData.c_str(), strData.size()); pData += strData.size();
+		*(WORD*)pData = sSum;
+		return strOut.c_str();
+	}
+
+	WORD sHead;//固定位FE FF
+	DWORD nLength;//包长度(包括下面3个!但是不包括自身);
+	WORD sCmd;//控制命令
+	std::string strData;//包数据，一个对象,传&仅仅会传一个地址，不是数据；
+	WORD sSum;	//和校验,只对strData求和
+	std::string strOut;
+};
+#pragma pack(pop)
+
+typedef struct MouseEvent {
+	MouseEvent() {
+		nAction = 0;
+		nButton = -1;
+		ptXY.x = 0;
+		ptXY.y = 0;
+	}
+	WORD nAction;//点击、移动、双击；
+	WORD nButton;//左键、右键、中键；
+	POINT ptXY;//坐标；
+}MOUSEEV, * PMOUSEEV;
+
+typedef struct file_info {
+	file_info() {
+		IsInvalid = FALSE;  //是否无效的文件
+		IsDirectory = -1;
+		hasNext = TRUE;
+		memset(szFileName, 0, sizeof(szFileName));
+	}
+	BOOL IsInvalid;
+	BOOL IsDirectory;
+	BOOL hasNext;
+	char szFileName[256];//文件名
+}FILEINFO, * PFILEINFO;
