+++ RemoteCtr/RemoteCtr/RemoteCtr.vcxproj
@@ -144,6 +144,7 @@
     <ClInclude Include="framework.h" />
     <ClInclude Include="LockDialog.h" />
     <ClInclude Include="MyTool.h" />
+    <ClInclude Include="Packet.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="RemoteCtr.h" />
     <ClInclude Include="Resource.h" />
