+++ RemoteCtr/RemoteCtr/RemoteCtr.vcxproj.filters
@@ -42,6 +42,9 @@
     <ClInclude Include="MyTool.h">
       <Filter>头文件</Filter>
     </ClInclude>
+    <ClInclude Include="Packet.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="RemoteCtr.cpp">
